+++ apps/account-server/src/app/_components/embedded.root.tsx
@@ -6,16 +6,14 @@ import {
   useRNHandler,
 } from '@sophon-labs/account-message-bridge';
 import { useCallback, useEffect, useState } from 'react';
-import { Button } from '@/components/ui/button';
 import { Drawer } from '@/components/ui/drawer';
 import { MainStateMachineContext } from '@/context/state-machine-context';
+import { logWithUser } from '@/debug/log';
 import { env } from '@/env';
-import { sendMessage } from '@/events';
 import { useEventHandler } from '@/events/hooks';
 import { useAccountContext } from '@/hooks/useAccountContext';
 import { useRequestDrawer } from '@/hooks/useRequestDrawer';
 import { useUserIdentification } from '@/hooks/useUserIdentification';
-import { serverLog } from '@/lib/server-log';
 import { CompletedView } from '@/views/CompletedView';
 import ConnectAuthorizationView from '@/views/ConnectAuthorizationView';
 import { LoadingView } from '@/views/LoadingView';
@@ -70,6 +68,7 @@ export default function EmbeddedRoot({ partnerId, scopes }: EmbeddedRootProps) {
     'openModal',
     useCallback(() => {
       setOpen(true);
+      logWithUser('From RN > opened auth modal');
     }, []),
   );
 
@@ -77,33 +76,48 @@ export default function EmbeddedRoot({ partnerId, scopes }: EmbeddedRootProps) {
     'closeModal',
     useCallback(() => {
       setOpen(false);
+      logWithUser('From RN > closed auth modal');
+    }, []),
+  );
+
+  useRNHandler(
+    'authSessionCancel',
+    useCallback(() => {
+      actorRef.send({ type: 'ACCOUNT_ERROR' });
+      logWithUser('From RN > cancelled auth session');
+    }, [actorRef]),
+  );
+
+  useRNHandler(
+    'authSessionRedirect',
+    useCallback((payload: { url: string }) => {
+      logWithUser(`From RN > url redirection to ${payload.url}`);
+      window.location.href = payload.url;
     }, []),
   );
 
   useRNHandler(
     'sdkStatusRequest',
     useCallback(() => {
-      sendMessageToRN('sdkStatusResponse', {
+      const payload = {
         isDrawerOpen: open,
         isReady: !state.context.isLoadingResources,
         isAuthenticated: state.context.isAuthenticated,
         connectedAccount: account?.address,
-      });
+      };
+      sendMessageToRN('sdkStatusResponse', payload);
+      logWithUser(`From RN > sdk status request ${JSON.stringify(payload)}`);
     }, [state, open, account?.address]),
   );
 
-  useEffect(() => {
-    serverLog(
-      `>>> 🔥 <<< STATE ${JSON.stringify(state.value)} / ${state.context.requests.incoming?.id}`,
-    );
-  }, [state]);
-
   useEventHandler('flow.complete', () => {
     setOpen(false);
+    logWithUser('flow complete');
   });
 
   useEventHandler('modal.open', () => {
     setOpen(true);
+    logWithUser('modal open');
   });
 
   const handleCloseModal = (isOpen: boolean) => {
@@ -287,30 +301,6 @@ export default function EmbeddedRoot({ partnerId, scopes }: EmbeddedRootProps) {
     );
   }
 
-  if (state.matches('profile') && account) {
-    const handleDisconnect = () => {
-      sendMessage('smart-contract.logout', null);
-      actorRef.send({ type: 'CANCEL' });
-    };
-
-    return (
-      <Drawer
-        open={open}
-        onOpenChange={handleCloseModal}
-        actions={<Button onClick={handleDisconnect}>Log out</Button>}
-        showHeader={true}
-        showProfileImage={true}
-        showLegalNotice={false}
-        showLogo={false}
-        title={account.address}
-        drawerType="user_profile"
-        // onSettings={() => {
-        //   // goToSettings();
-        // }}
-      />
-    );
-  }
-
   if (state.matches('completed')) {
     return (
       <Drawer
@@ -325,40 +315,6 @@ export default function EmbeddedRoot({ partnerId, scopes }: EmbeddedRootProps) {
       </Drawer>
     );
   }
-
-  // TODO: settings state
-  // if (authState === AuthState.SETTINGS) {
-  //   return (
-  //     <Drawer
-  //       open={open}
-  //       onOpenChange={(open) => {
-  //         setOpen(open);
-  //         if (!open) {
-  //           windowService.close();
-  //         }
-  //       }}
-  //       showHeader={true}
-  //       showLegalNotice={false}
-  //       showLogo={false}
-  //       title="Settings"
-  //       actions={
-  //         <div className="flex flex-col gap-2 w-full items-center">
-  //           <p className="text-lg">Manage your account at Sophon Home</p>
-  //           <Button
-  //             onClick={() => {
-  //               window.parent.open('https://app.sophon.xyz/', '_blank');
-  //             }}
-  //           >
-  //             Open Sophon
-  //           </Button>
-  //         </div>
-  //       }
-  //       onBack={() => {
-  //         goToAuthenticated();
-  //       }}
-  //     />
-  //   );
-  // }
 
   return (
     <Drawer