+++ apps/account-server/src/hooks/useMessageHandler.ts
@@ -1,6 +1,7 @@
 'use client';
 import { useEffect, useState } from 'react';
 import { hexToString, toHex } from 'viem';
+import { logWithUser } from '@/debug/log';
 import { isValidPaymaster } from '@/lib/paymaster';
 import { windowService } from '@/service/window.service';
 import type {
@@ -52,6 +53,7 @@ export const useMessageHandler = (): UseMessageHandlerReturn => {
       // Store the incoming request if it's an RPC request
       if (data?.id && data?.content) {
         const method = data.content?.action?.method;
+        logWithUser(`RPC Method Requested > ${method} > ${data.id}`);
         if (method === 'eth_requestAccounts') {
           const params = data.content.action?.params as
             | { sessionPreferences?: unknown }