+++ packages/account-react-native/src/hooks/use-sophon-account.ts
@@ -1,24 +1,33 @@
 import { useCallback, useMemo, useState } from 'react';
 import type { Address } from 'viem';
 import type { CustomRPCError } from '@/types';
+import { sendUIMessage } from '../messaging';
 import { useSophonContext } from './use-sophon-context';
 
 export const useSophonAccount = () => {
   const {
+    initialized,
     walletClient,
     setAccount,
     provider,
     account,
-    disconnect,
     error,
     logout,
+    disconnect,
   } = useSophonContext();
   const [accountError, setAccountError] = useState<string>();
   const [isConnecting, setIsConnecting] = useState(false);
 
   const connect = useCallback(async () => {
+    setIsConnecting(true);
+
+    // to make sure that we have no cached account, before connecting we force a local disconnect
     try {
-      setIsConnecting(true);
+      await disconnect();
+      sendUIMessage('clearMainViewCache', {});
+    } catch {}
+
+    try {
       setAccountError(undefined);
       const addresses = await walletClient!.requestAddresses();
       if (addresses.length === 0) {
@@ -37,12 +46,18 @@ export const useSophonAccount = () => {
     }
   }, [walletClient, setAccount]);
 
-  const isConnected = useMemo(() => !!account, [account]);
+  // Make sure to only return that the user is connected after
+  // context initialization is complete, that way we make sure that
+  // we don't load  cached stuff from previous installation
+  const isConnected = useMemo(
+    () => !!account && initialized,
+    [account, initialized],
+  );
 
   return {
+    initialized,
     isConnected,
     connect,
-    disconnect,
     logout,
     account,
     provider,